+++ zio-schema-examples/shared/src/main/scala/dev/zio/schema/example/example1/Example1.scala
@@ -185,8 +185,7 @@ object ProtobufExample extends ZIOAppDefault {
 
 object CombiningExample extends ZIOAppDefault {
   import ManualConstruction._
-  import zio.schema.codec.{ ProtobufCodec }
-  import zio.schema.codec.JsonCodec
+  import zio.schema.codec.{ JsonCodec, ProtobufCodec }
   import zio.stream.ZStream
 
   override def run: ZIO[Environment with ZIOAppArgs, Any, Any] =
